Started working on making the js data files to json format again. This was a mistake i made in the beginning of this project, so im trying to make it with json files again. Atleast its good practicegit add .
