+++ src/server/data/about.json
@@ -1,0 +1,22 @@
+{
+  "aboutUs": "Om oss",
+  "headline": "Om KINO Sandviken",
+  "aboutPage": [
+    {
+      "section": "Om Företaget",
+      "content": "Sedan 2023 har Kino Bio varit stadens biograf för filmälskare. Här möts klassisk charm och modern teknik i våra toppmoderna salonger. Vi visar allt från storfilmer till indiefavoriter och bjuder på en unik bioupplevelse för alla. Kom och upplev filmens värld hos oss – din berättelse börjar här!"
+    },
+    {
+      "section": "Vi Gör Bioupplevelsen Bättre",
+      "content": "På Kino Bio är vårt löfte att alltid erbjuda en bioupplevelse utöver det vanliga. Vi strävar efter att skapa minnesvärda stunder och förstklassig underhållning som gör varje besök unikt och oförglömligt."
+    },
+    {
+      "section": "Upplev Bubbelpoolsbio på Kino Bio",
+      "content": "Hos oss tar vi bioupplevelsen till nästa nivå med vår exklusiva bubbelpoolsbio. Koppla av i en varm bubbelpool samtidigt som du njuter av de senaste filmerna på en storslagen skärm. Perfekt för en lyxig filmkväll med vänner eller en romantisk stund. Kombinera avslappning och underhållning på ett sätt du aldrig upplevt tidigare!"
+    },
+    {
+      "section": "Vår Webbplats och App",
+      "content": "Kinos app kan du enkelt ladda ner på App Store eller Google Play."
+    }
+  ]
+}
